Redesigned the struture of the SpotifyAPIDataAccessObject to make it store playlists info to the playlist csv file

+++ src/use_case/editProfile/SpotifyAPIDataAccessInterface.java
@@ -10,13 +10,13 @@ public interface SpotifyAPIDataAccessInterface {
 
     // getAcousticness(Playlist.playlistId)
     // 0-1
-    float getAcousticness(String playlistId);
+    double getAcousticness(String playlistId);
     // 0-1
-    float getEnergy(String playlistId);
+    double getEnergy(String playlistId);
     // 0-1
-    float getInstrumentalness(String playlistId);
+    double getInstrumentalness(String playlistId);
     // 0-1
-    float getValence(String playlistId);
+    double getValence(String playlistId);
     // String
     String getGenre(String playlistId);
